// Client/src/data/categories.ts
export interface Category {
  id: string;
  name: string;
  description: string;
  icon: string;
}

// export const initialCategories: Category[] = [
// {
//   id: '1',
//   name: 'IT Equipment',
//   description: 'Computers, laptops, servers, networking equipment',
//   icon: 'Monitor',
// },
// {
//   id: '2',
//   name: 'Office Furniture',
//   description: 'Desks, chairs, cabinets, tables',
//   icon: 'Armchair',
// },
// {
//   id: '3',
//   name: 'Tools',
//   description: 'Power tools, hand tools, maintenance equipment',
//   icon: 'Wrench',
// },
// ];